requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiolimiter==1.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==21.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fonttools==4.55.0
frozenlist==1.5.0
grobid-client==0.8.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lmdb==1.5.1
lxml==4.9.4
matplotlib==3.9.2
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.1.3
ollama==0.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pycurl==7.45.3
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
strsimpy==0.2.1
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
wayback==0.4.5
xloop==1.1.0
xsentinels==1.2.2
yarl==1.18.3
